Refactor methods in game_stats and create team utility module
