+++ spec/spec_helper.rb
@@ -17,3 +17,4 @@ require './lib/team'
 require './lib/statistics_module'
 require './lib/season_utility'
 require './lib/goals_utility'
+require './team_utility'
